move code for allowing vdf praams to be stored instead  of hard coded.

+++ types/src/ol_vdf_difficulty.rs
@@ -20,7 +20,7 @@ pub struct VDFDifficulty {
     /// Difficulty
     pub difficulty: u64,
     /// Security parameter for VDF
-    pub security: u16,
+    pub security: u64,
 }
 
 
@@ -31,7 +31,7 @@ impl VDFDifficulty {
     }
 
     /// get the security param of the block, or assume legacy
-    pub fn security(&self) -> u16 {
+    pub fn security(&self) -> u64 {
       self.security
     }
